Divide the header file
