--- include/minishell_tfujiwar.h
@@ -1,16 +1,0 @@
-# define STATUS_BASE			128
-
-# define MSG_SIG_QUIT			"Quit: 3"
-# define MSG_SYNTAX_ERR			"syntax error"
-# define MSG_HEREDOC_EOF		"here-document delimited end-of-file"
-
-# define PROMPT_MINISH			"minishell $ "
-# define PROMPT_HREDOC			"> "
-
-# define END_OF_IO				-1
-# define INVALID_FD				-1
-
-# define CHRS_PLAIN_DELIM		"\'\"$"
-
-# define LEN_QUOTE_CLOSED		2
-# define LEN_QUOTE_UNCLOSED		1
