+++ include/minishell_typedef.h
@@ -1,0 +1,63 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   minishell_typedef.h                                :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: Yoshihiro Kosaka <ykosaka@student.42tok    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/12/01 13:51:23 by tfujiwar          #+#    #+#             */
+/*   Updated: 2022/12/31 17:08:15 by Yoshihiro K      ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef MINISHELL_TYPEDEF_H
+# define MINISHELL_TYPEDEF_H
+
+# include <unistd.h>
+# include <stdio.h>
+# include <stdlib.h>
+# include <fcntl.h>
+# include <sys/types.h>
+# include <sys/stat.h>
+# include <dirent.h>
+# include <signal.h>
+# include <sys/wait.h>
+# include <readline/readline.h>
+# include <readline/history.h>
+# include <stdbool.h>
+# include <limits.h>
+# include <errno.h>
+# include <string.h>
+# include "libft.h"
+
+typedef struct sigaction	t_sa;
+
+typedef struct s_token {
+	char	*str;
+	int		flag;
+}	t_token;
+
+typedef struct s_fd {
+	char	*path;
+	int		fd;
+}	t_fd;
+
+typedef struct s_cmd {
+	char			*path;
+	char			**arg;
+	t_fd			*input;
+	t_fd			*output;
+	struct s_cmd	*next;
+	struct s_cmd	*prev;
+	pid_t			pid;
+}	t_cmd;
+
+typedef struct s_shell {
+	t_list	*environ;
+	int		status;
+	t_cmd	*cmd;
+	bool	kill_child;
+	bool	heredoc_sigint;
+}	t_shell;
+
+#endif
